refactor: move MultiSelect to components/ui/ per convention
Follow project convention: generic UI primitives go in src/components/ui/
alongside other headless components (Popover, Command, Select).
